+++ internal/provider/resource_iosxe_evpn_instance_test.go
@@ -39,12 +39,6 @@ func TestAccIosxeEVPNInstance(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_evpn_instance.test", "vlan_based_encapsulation", "vxlan"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_evpn_instance.test", "vlan_based_auto_route_target", "false"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_evpn_instance.test", "vlan_based_rd", "10:10"))
-	if os.Getenv("IOSXE179") != "" || os.Getenv("IOSXE1712") != "" {
-		checks = append(checks, resource.TestCheckResourceAttr("iosxe_evpn_instance.test", "vlan_based_route_target", "10:10"))
-	}
-	if os.Getenv("IOSXE179") != "" || os.Getenv("IOSXE1712") != "" {
-		checks = append(checks, resource.TestCheckResourceAttr("iosxe_evpn_instance.test", "vlan_based_route_target_both", "10:10"))
-	}
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_evpn_instance.test", "vlan_based_route_target_import", "10:10"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_evpn_instance.test", "vlan_based_route_target_export", "10:10"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_evpn_instance.test", "vlan_based_ip_local_learning_disable", "false"))
@@ -88,12 +82,6 @@ func testAccIosxeEVPNInstanceConfig_all() string {
 	config += `	vlan_based_encapsulation = "vxlan"` + "\n"
 	config += `	vlan_based_auto_route_target = false` + "\n"
 	config += `	vlan_based_rd = "10:10"` + "\n"
-	if os.Getenv("IOSXE179") != "" || os.Getenv("IOSXE1712") != "" {
-		config += `	vlan_based_route_target = "10:10"` + "\n"
-	}
-	if os.Getenv("IOSXE179") != "" || os.Getenv("IOSXE1712") != "" {
-		config += `	vlan_based_route_target_both = "10:10"` + "\n"
-	}
 	config += `	vlan_based_route_target_import = "10:10"` + "\n"
 	config += `	vlan_based_route_target_export = "10:10"` + "\n"
 	config += `	vlan_based_ip_local_learning_disable = false` + "\n"