+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.6.0 (unrelease)
+
+- Drop support for IOS-XE 17.9
+- Add `esp-192-aes` and `esp-256-aes` options to `iosxe_crypto_ipsec_transform_set` resource and data source
+- Make `v3_auth_algorithm` attribute of `iosxe_snmp_server_user` resource and data source optional
+- Add `TLSv1.3` option to `ip_http_tls_version` attribute of `iosxe_system` resource and data source
+
 ## 0.5.10
 
 - Fix update issue with `iosxe_restconf` resource and empty `attributes` map